% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.quantileGA.R
\name{plot.quantileGA}
\alias{plot.quantileGA}
\title{Plot diagnostics for a \code{quantileGA} model.}
\usage{
\method{plot}{quantileGA}(x, type = c("convergence", "residuals"), ...)
}
\arguments{
\item{x}{\code{quantileGA}\cr Object returned by \code{GA.QR}.}

\item{type}{\code{character}\cr Type of plot. Either \code{"convergence"}
to show the evolution of the fitness or \code{"residuals"} to plot
residuals. Defaults to \code{"convergence"}.}

\item{...}{Additional graphical parameters passed to \code{plot}.}
}
\value{
Invisibly returns \code{NULL}. A plot is produced as a side effect.
}
\description{
Plot diagnostics for a \code{quantileGA} model.
}
