% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.quantileGA.R
\name{print.quantileGA}
\alias{print.quantileGA}
\title{Print method for \code{quantileGA} objects.}
\usage{
\method{print}{quantileGA}(x, ...)
}
\arguments{
\item{x}{\code{quantileGA}\cr Object produced by \code{GA.QR}.}

\item{...}{Further arguments passed to or from other methods. Currently
ignored.}
}
\value{
Invisibly returns a matrix of coefficients and related statistics.
}
\description{
Print method for \code{quantileGA} objects.
}
