% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantGA-package.R
\docType{package}
\name{quantGA}
\alias{quantGA-package}
\alias{quantGA}
\title{quantGA: Fits a linear quantile regression model using genetic algorithm}
\description{
Provides tools to fit linear quantile regression models using a genetic
algorithm to estimate the coefficients.
}
\author{
\strong{Maintainer}: Adrien Orue \email{adrienorue@example.com}

}
\keyword{internal}
