% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.quantileGA.R
\name{summary.quantileGA}
\alias{summary.quantileGA}
\title{Summary method for \code{quantileGA} objects.}
\usage{
\method{summary}{quantileGA}(object, ...)
}
\arguments{
\item{object}{\code{quantileGA}\cr Object returned by \code{GA.QR}.}

\item{...}{Additional arguments passed to or from other methods. Currently
ignored.}
}
\value{
Invisibly returns \code{object} after printing a short summary.
}
\description{
Summary method for \code{quantileGA} objects.
}
